feat: Add Ranges filter type
Numeric range filters can be displayed as a slider or inputs,
in addition to the default list of options display.
A new question type "Range" of Celebros API is supported.
A dynamic property “RangeDisplay” defines a visual representation
of a Range filter. If the value of that dynamic property is “Slider”
it displays a slider for this filter. If the value is “Inputs”
it displays min and max value inputs for this filter.
For any other value, or if the dynamic property doesn’t exist,
it displays a regular filter.

ref: SSUITE-917